requirements.txt
requests
tqdm
PyPDF2
python-docx
fastapi
uvicorn
chromadb
pyyaml
watchdog
